V3/backend/requirements.txt
# Core runtime
opencv-python==4.10.0.84
mediapipe>=0.10.14
numpy>=1.26
pillow>=10.3.0
tqdm>=4.66.4

# Windows integration & input
pywin32>=306
psutil>=5.9.8
pydirectinput>=1.0.4
keyboard>=0.13.5
mouse>=0.7.1

# Local API (optional) — not a webapp UI, just an API you can call from a native UI later
fastapi>=0.112.1
uvicorn[standard]>=0.30.5
pydantic>=2.8.2

# OpenAI Vision assist (optional)
openai>=1.51.0
PyQt6>=6.6.1
python-dotenv>=1.0.1
requests>=2.32.3